First tests

+++ src/main/java/org/example/Boot.java
@@ -2,9 +2,9 @@ package org.example;
 
 public class Boot extends Item{
     private BootType type;
-    private float size;
+    private double size;
 
-    public Boot(String name, Producent producent, int discountValue, float price, AdvancementLevel advancementLevel, BootType type, float size) {
+    public Boot(String name, Producent producent, int discountValue, double price, AdvancementLevel advancementLevel, BootType type, double size) {
         super(name, producent, discountValue, price, advancementLevel);
         this.type = type;
         this.size = size;
